+++ common-libs/internal-api-clients/src/spec/_keyStorage.ts
@@ -153,6 +153,136 @@ export default {
 				"type": "object",
 				"additionalProperties": false
 			},
+			"ProfileTrueCaller": {
+				"properties": {
+					"payload": {
+						"type": "string"
+					},
+					"signature": {
+						"type": "string"
+					},
+					"signatureAlgorithm": {
+						"type": "string"
+					},
+					"avatarUrl": {
+						"type": "string",
+						"nullable": true
+					},
+					"city": {
+						"type": "string",
+						"nullable": true
+					},
+					"companyName": {
+						"type": "string",
+						"nullable": true
+					},
+					"countryCode": {
+						"type": "string",
+						"nullable": true
+					},
+					"email": {
+						"type": "string",
+						"nullable": true
+					},
+					"facebookId": {
+						"type": "string",
+						"nullable": true
+					},
+					"firstName": {
+						"type": "string"
+					},
+					"gender": {
+						"type": "string"
+					},
+					"isAmbassador": {
+						"type": "boolean"
+					},
+					"isBusiness": {
+						"type": "boolean"
+					},
+					"isVerified": {
+						"type": "boolean"
+					},
+					"jobTitle": {
+						"type": "string",
+						"nullable": true
+					},
+					"lastName": {
+						"type": "string"
+					},
+					"phoneNumber": {
+						"type": "string"
+					},
+					"requestNonce": {
+						"type": "string"
+					},
+					"street": {
+						"type": "string",
+						"nullable": true
+					},
+					"successful": {
+						"type": "boolean"
+					},
+					"twitterId": {
+						"type": "string",
+						"nullable": true
+					},
+					"url": {
+						"type": "string",
+						"nullable": true
+					},
+					"zipcode": {
+						"type": "string",
+						"nullable": true
+					}
+				},
+				"required": [
+					"payload",
+					"signature",
+					"signatureAlgorithm"
+				],
+				"type": "object",
+				"additionalProperties": {}
+			},
+			"TruecallerUserListInput": {
+				"properties": {
+					"profileTrueCaller": {
+						"$ref": "#/components/schemas/ProfileTrueCaller"
+					},
+					"expiredAt": {
+						"type": "string"
+					},
+					"env": {
+						"type": "string",
+						"nullable": true
+					}
+				},
+				"required": [
+					"profileTrueCaller"
+				],
+				"type": "object",
+				"additionalProperties": false
+			},
+			"AdminCreateUserInput": {
+				"properties": {
+					"profileTrueCaller": {
+						"$ref": "#/components/schemas/ProfileTrueCaller"
+					},
+					"password": {
+						"type": "string"
+					},
+					"username": {
+						"type": "string"
+					}
+				},
+				"required": [
+					"profileTrueCaller",
+					"password",
+					"username"
+				],
+				"type": "object",
+				"additionalProperties": false
+			},
 			"AdminConfirmUserInput": {
 				"properties": {
 					"username": {
@@ -215,6 +345,36 @@ export default {
 				"responses": {
 					"204": {
 						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "Store custom message template for passwordless auth flow.",
@@ -241,6 +401,36 @@ export default {
 				"responses": {
 					"204": {
 						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "Delete custom message template for passwordless auth flow.",
@@ -271,6 +461,36 @@ export default {
 							"application/json": {
 								"schema": {
 									"$ref": "#/components/schemas/GetCredentialOfferOutput"
+								}
+							}
+						}
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
 								}
 							}
 						}
@@ -326,6 +546,36 @@ export default {
 								}
 							}
 						}
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "Route used by SDK for getting users email & phone number signed credentials.",
@@ -372,6 +622,36 @@ export default {
 								}
 							}
 						}
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "Read my encrypted key from a key storage",
@@ -408,6 +688,46 @@ export default {
 								}
 							}
 						}
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"409": {
+						"description": "Key for this user already exists",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "Store my encrypted key on a key storage",
@@ -441,12 +761,192 @@ export default {
 				}
 			}
 		},
+		"/truecaller/storeUserInList": {
+			"post": {
+				"operationId": "StoreInTruecallerUserList",
+				"responses": {
+					"204": {
+						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					}
+				},
+				"description": "Store `Truecaller` user in DDB for further use in Cognito lambdas.",
+				"tags": [
+					"StoreTruecallerUserInList"
+				],
+				"security": [
+					{
+						"truecallerAuth": []
+					}
+				],
+				"parameters": [],
+				"requestBody": {
+					"required": true,
+					"content": {
+						"application/json": {
+							"schema": {
+								"$ref": "#/components/schemas/TruecallerUserListInput"
+							}
+						}
+					}
+				}
+			}
+		},
+		"/userManagement/adminCreateConfirmedUser": {
+			"post": {
+				"operationId": "AdminCreateConfirmedUser",
+				"responses": {
+					"204": {
+						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"409": {
+						"description": "User {{username}} already exists.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Missing configuration for AWS in environment variables.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					}
+				},
+				"description": "This endpoint should be used to create user confirmed user with a `Truecaller` profile(token).",
+				"tags": [
+					"UserManagement"
+				],
+				"security": [
+					{
+						"truecallerAuth": []
+					}
+				],
+				"parameters": [],
+				"requestBody": {
+					"required": true,
+					"content": {
+						"application/json": {
+							"schema": {
+								"$ref": "#/components/schemas/AdminCreateUserInput"
+							}
+						}
+					}
+				}
+			}
+		},
 		"/userManagement/adminConfirmUser": {
 			"post": {
 				"operationId": "AdminConfirmUser",
 				"responses": {
 					"204": {
 						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"404": {
+						"description": "User {{username}} not found.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"409": {
+						"description": "User {{username}} already confirmed.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Missing configuration for AWS in environment variables.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "This endpoint should be used to confirm user creation.",
@@ -473,6 +973,66 @@ export default {
 				"responses": {
 					"204": {
 						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"404": {
+						"description": "User {{username}} not found.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"409": {
+						"description": "User {{username}} must be UNCONFIRMED (current status is {{userStatus}}).",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Missing configuration for AWS in environment variables.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "This endpoint should be used to delete unconfirmed user.",
@@ -499,6 +1059,46 @@ export default {
 				"responses": {
 					"204": {
 						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Missing configuration for AWS in environment variables.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "This endpoint should be used as a fallback in case user creation hangs,\nand it should delete the user from cognito, to allow retries.",
@@ -528,6 +1128,46 @@ export default {
 				"responses": {
 					"204": {
 						"description": "No content"
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Missing configuration for AWS in environment variables.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
 					}
 				},
 				"description": "Signs out a user from all devices.\nIt also invalidates all refresh tokens that Amazon Cognito has issued to a user.\nThe user's current access and ID tokens remain valid until they expire.\nBy default, access and ID tokens expire one hour after they're issued.\nA user can still use a hosted UI cookie to retrieve new tokens for the duration of the cookie validity period of 1 hour.",
@@ -600,6 +1240,46 @@ export default {
 										"userExists"
 									],
 									"type": "object"
+								}
+							}
+						}
+					},
+					"400": {
+						"description": "Bad Request",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"401": {
+						"description": "Unauthorized",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"405": {
+						"description": "Method Not Allowed",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
+								}
+							}
+						}
+					},
+					"422": {
+						"description": "Missing configuration for AWS in environment variables.",
+						"content": {
+							"application/json": {
+								"schema": {
+									"$ref": "#/components/schemas/ErrorObject"
 								}
 							}
 						}